+++ crates/objc2/src/macros/extern_class.rs
@@ -337,6 +337,13 @@ macro_rules! __inner_extern_class {
             }
         }
 
+        // SAFETY: This maps `SomeClass<T, ...>` to a single `SomeClass<AnyObject, ...>` type and
+        // implements `DowncastTarget` on that type. This is safe because the "base container" class
+        // is the same and each generic argument is replaced with `AnyObject`, which can represent
+        // any Objective-C class instance.
+        $(#[$impl_m])*
+        unsafe impl $crate::DowncastTarget for $name<$($crate::__extern_class_map_anyobject!($t_for)),*> {}
+
         $(#[$impl_m])*
         unsafe impl<$($t_for $(: $(?$b_sized_for +)? $b_for)?),*> ClassType for $for {
             type Super = $superclass;
@@ -357,6 +364,14 @@ macro_rules! __inner_extern_class {
             #[inline]
             fn as_super(&$as_super_self) -> &Self::Super $as_super
         }
+    };
+}
+
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __extern_class_map_anyobject {
+    ($t:ident) => {
+        $crate::runtime::AnyObject
     };
 }