+++ framework-crates/objc2-foundation/src/tests/string.rs
@@ -80,17 +80,17 @@ fn test_interior_nul() {
 #[cfg(feature = "NSObject")]
 fn test_copy() {
     use crate::{NSCopying, NSMutableCopying, NSMutableString, NSObjectProtocol};
-    use objc2::rc::Retained;
+    use objc2::{rc::Retained, ClassType};
 
     let s1 = NSString::from_str("abc");
     let s2 = s1.copy();
     // An optimization that NSString makes, since it is immutable
     assert_eq!(Retained::as_ptr(&s1), Retained::as_ptr(&s2));
-    assert!(s2.is_kind_of::<NSString>());
+    assert!(s2.isKindOfClass(NSString::class()));
 
     let s3 = s1.mutableCopy();
     assert_ne!(Retained::as_ptr(&s1), Retained::as_ptr(&s3).cast());
-    assert!(s3.is_kind_of::<NSMutableString>());
+    assert!(s3.isKindOfClass(NSMutableString::class()));
 }
 
 #[test]